Refactoring for data types decoupling
