+++ app/models/setup/flow.rb
@@ -15,7 +15,7 @@ module Setup
     belongs_to :event, class_name: Setup::Event.to_s, inverse_of: nil
 
     belongs_to :translator, class_name: Setup::Translator.to_s, inverse_of: nil
-    belongs_to :custom_data_type, class_name: Setup::Model.to_s, inverse_of: nil
+    belongs_to :custom_data_type, class_name: Setup::DataType.to_s, inverse_of: nil
     field :nil_data_type, type: Boolean
     field :data_type_scope, type: String
     field :scope_filter, type: String
@@ -25,7 +25,7 @@ module Setup
     belongs_to :connection_role, class_name: Setup::ConnectionRole.to_s, inverse_of: nil
 
     belongs_to :response_translator, class_name: Setup::Translator.to_s, inverse_of: nil
-    belongs_to :response_data_type, class_name: Setup::Model.to_s, inverse_of: nil
+    belongs_to :response_data_type, class_name: Setup::DataType.to_s, inverse_of: nil
 
     field :last_trigger_timestamps, type: Time
