+++ app/models/setup/observer.rb
@@ -4,7 +4,7 @@ module Setup
 
     BuildInDataType.regist(self).referenced_by(:name).excluding(:last_trigger_timestamps).including(:data_type)
 
-    belongs_to :data_type, class_name: Setup::Model.to_s, inverse_of: :events
+    belongs_to :data_type, class_name: Setup::DataType.to_s, inverse_of: :events
     field :triggers, type: String
 
     validates_presence_of :data_type, :triggers
