+++ config/application.rb
@@ -29,22 +29,21 @@ module Cenit
     config.after_initialize do
 
       model_update_options = {model_loaded: false, used_memory: 0}
-      model_update_options[:activated] = false if Cenit.deactivate_models
+      model_update_options[:activated] = false #if Cenit.deactivate_models
 
       Account.all.each do |account|
 
         Account.current = account
 
-        Setup::Model.update_all(model_update_options)
+        Setup::DataType.update_all(model_update_options)
 
         unless Cenit.deactivate_models
-          Setup::Schema.all.each do |schema|
-            puts "Loading schema #{schema.uri}"
-            schema.load_models
+          models = Set.new
+          Setup::SchemaDataType.activated.each do |data_type|
+            models += data_type.load_models[:loaded]
           end
-          models = []
-          Setup::FileDataType.all.each do |file_data_type|
-            models << file_data_type.load_model if file_data_type.activated
+          Setup::FileDataType.activated.each do |file_data_type|
+            models << file_data_type.load_model
           end
           RailsAdmin::AbstractModel.update_model_config(models)
         end
