+++ lib/rails_admin/config/actions/convert.rb
@@ -18,7 +18,7 @@ module RailsAdmin
         end
 
         register_instance_option :except do
-          [Setup::Library, Setup::Schema, Setup::Model]
+          [Setup::Library, Setup::Schema, Setup::DataType]
         end
 
         register_instance_option :link_icon do
