+++ lib/rails_admin/config/actions/data_type.rb
@@ -24,7 +24,7 @@ module RailsAdmin
           proc do
 
             if data_type = abstract_model.model_name.constantize.try(:data_type)
-              redirect_to rails_admin.show_path(model_name: Setup::Model.to_s.underscore.gsub('/', '~'), id: data_type.id.to_s)
+              redirect_to rails_admin.show_path(model_name: data_type.class.to_s.underscore.gsub('/', '~'), id: data_type.id.to_s)
             else
               redirect_to back_or_index
             end
