+++ lib/rails_admin/config/actions/shutdown_model.rb
@@ -4,7 +4,7 @@ module RailsAdmin
       class ShutdownModel < RailsAdmin::Config::Actions::Base
 
         register_instance_option :only do
-          Setup::Model
+          [Setup::DataType, Setup::SchemaDataType, Setup::FileDataType]
         end
 
         register_instance_option :member do
@@ -25,14 +25,14 @@ module RailsAdmin
             if @object.activated && @object.loaded?
               begin
                 unless params[:shutdown]
-                  report = Setup::Model.shutdown(@object, report_only: true)
+                  report = Setup::DataType.shutdown(@object, report_only: true)
                   @shutdown = report[:destroyed].collect(&:data_type).uniq.select { |data_type| data_type != @object }
                   @object.instance_variable_set(:@_to_reload, reload = report[:affected].collect(&:data_type).uniq)
                   params[:shutdown] = true if @shutdown.empty? && reload.empty?
                 end
                 if params[:shutdown]
                   @object.activated = ok = false
-                  (report = Setup::Model.shutdown(@object))[:destroyed].each do |model|
+                  (report = Setup::DataType.shutdown(@object))[:destroyed].each do |model|
                     (data_type = model.data_type).activated = data_type.show_navigation_link = false
                     data_type.save
                     ok = ok || data_type == @object
@@ -55,7 +55,7 @@ module RailsAdmin
             else
               flash[:error] = "Can not explicitly shutdown model '#{@object.title}'"
             end
-            redirect_to rails_admin.show_path(model_name: Setup::Model.to_s.underscore.gsub('/', '~'), id: @object.id) if done
+            redirect_to rails_admin.show_path(model_name: Setup::DataType.to_s.underscore.gsub('/', '~'), id: @object.id) if done
           end
         end
