+++ lib/rails_admin/config/fields/types/json_schema.rb
@@ -4,6 +4,11 @@ module RailsAdmin
       module Types
         class JsonSchema < RailsAdmin::Config::Fields::Types::CodeMirror
           include JsonValueCommon
+
+          def parse_input(params)
+            super
+            params[name] = nil unless params[name].is_a?(Hash)
+          end
         end
       end
     end
