+++ spec/factories/webhooks.rb
@@ -3,7 +3,7 @@ FactoryGirl.define do
     name 'Add Product'
     path 'add_product'
     purpose 'send'
-    data_type { Setup::DataType.where(name: 'Product').first }
+    data_type { Setup::SchemaDataType.where(name: 'Product').first }
     
     # ensure product_schema and product_data_type will be created
     before(:create) { create(:product_schema) if Setup::Schema.where(uri: 'Product').count == 0 }
@@ -13,7 +13,7 @@ FactoryGirl.define do
     name 'Update Product'
     path 'update_product'
     purpose 'send'
-    data_type { Setup::DataType.where(name: 'Product').first }
+    data_type { Setup::SchemaDataType.where(name: 'Product').first }
     
     # ensure product_schema and product_data_type will be created
     before(:create) { create(:product_schema) if Setup::Schema.where(uri: 'Product').count == 0 }
